requirements.txt
tweepy
schedule
pymongo
dnspython
python-dotenv
black